requirements.txt
GitPython
mlflow
pytorch_lightning
PyYAML
hydra-core
tqdm
transformers
setuptools
sentencepiece
openpyxl
pandas
black
lint
pipreqs
sklearn
torchvision
datasets